Add authorization checking for meal plans

+++ src/services/mealPlans.ts
@@ -1,5 +1,6 @@
 import { mealPlansRepository } from "../dataaccess/mealPlans";
 import { CreateMealPlanSchema, UpdateMealPlanSchema } from "../dto/mealPlans";
+import { NotFoundError, UnauthorizedError } from "../errors/errors";
 import { MealPlan, MealPlanWithNutrition, Nutrition } from "../types";
 import { sumMealNutrition } from "./nutrition";
 
@@ -46,17 +47,33 @@ async function getMealPlan(mealPlanId: number) {
   return withNutrition(mealPlan);
 }
 
-async function createMealPlan(schema: CreateMealPlanSchema) {
-  const mealPlan = await mealPlansRepository.createMealPlan(schema);
+async function createMealPlan(schema: CreateMealPlanSchema, userId: number) {
+  const mealPlan = await mealPlansRepository.createMealPlan(schema, userId);
   return withNutrition(mealPlan);
 }
 
-async function updateMealPlan(schema: UpdateMealPlanSchema) {
-  const mealPlan = await mealPlansRepository.updateMealPlan(schema);
-  return withNutrition(mealPlan);
+async function updateMealPlan(schema: UpdateMealPlanSchema, userId: number) {
+  const mealPlan = await mealPlansRepository.getMealPlan(schema.id);
+  if (!mealPlan) {
+    throw new NotFoundError();
+  }
+  // Ensure that the meal plan can only be updated by its creator
+  if (mealPlan.creatorId !== userId) {
+    throw new UnauthorizedError();
+  }
+  const updatedMealPlan = await mealPlansRepository.updateMealPlan(schema);
+  return withNutrition(updatedMealPlan);
 }
 
-async function deleteMealPlan(mealPlanId: number) {
+async function deleteMealPlan(mealPlanId: number, userId: number) {
+  const mealPlan = await mealPlansRepository.getMealPlan(mealPlanId);
+  if (!mealPlan) {
+    throw new NotFoundError();
+  }
+  // Ensure that the meal plan can only be deleted by its creator
+  if (mealPlan.creatorId !== userId) {
+    throw new UnauthorizedError();
+  }
   return mealPlansRepository.deleteMealPlan(mealPlanId);
 }
